Clean up plugins repository
Some of the column names are too long.

The actual name of the folder where the plugin is installed
really doesn't matter, just the name inside the plugin.info file.
Also, by including a specific commit, we should ensure that a
plugin developer will have to re-submit their plugin whenever
they want to commit a new version, which means that the Naomi
devs will have a heads up that the plugin needs to be re-examined
each time.