Pocketsphinx_KWS README.md fix
Updated to reference the version of Pocketsphinx_KWS which has the
lower case readme.md file.

Also merged some changes to the command-line output of the
NaomiSTTTrainer.py server when importing audiolog data so you
can see that the server is doing things and hasn't just crashed.
Updated the commit version of this plugin also.