Added the Naomi2 TTI plugin
The Naomi2 TTI plugin is a simplified version of the Naomi TTI
intent parser. This version tries to use word frequencies to match
a specific template rather than to match a whole intent (which
contains many templates which may or may not all relate to the
same actual intention of the user).

For example, the old Naomi TTI had problems with templates that
were very similar to each other. The MPDControl command
"play <playlist>" and the Frotz command "play <game>" are very
similar, but in practive "play <game>" would always win even
though <playlist> is only associated with MPDControl just because
MPDControl has a bunch of other templates like "increase volume"
and "pause" which did not include the word "play" while every
template for Frotz did include the word "play".

"Tell me a joke" would also invariably choose a knock knock joke
since it had the only template that matched all of those words.

The new system will use Frotz to play a game or MPDControl to
play a playlist, and will choose a joke at random from the
different intents that include the words "tell me a ... joke".

Any feedback greatly appreciated.