Update Pocketsphinx_KWS
Pocketsphinx_KWS now uses the Python bindings for Phonetisaurus.
It also now can download and install pocketsphinx. In the process
of creating this update I fixed and improved this plugin and
discovered that it does make a difference what the threshold is
set to. I have kept the default of 10e-30. You can set the
threshold for a specific keyword in the profile. I hope to come
out with a new NaomiSTTTrainer.py plugin to use audiolog data to
optimize this value.