Update FAISS TTI to version 0.2.0
This updated version contains a few bug fixes. First, matches work
with some coveats now, but should work in pretty much any current
uses. As long as the Keywords in the Templates are associated with
lists of words, it should work.

Also updated the python_requirements files to require faiss-cpu
instead of just faiss.